By pass filterEverything for @api and @track properties
The benchmark tests are setting table data as an @api value. Locker's current POC implementation filters the array values. It is a know issue with locker when it comes to filtering arrays. 
For the purpose of this test, we will skip filtering @api and @track properties because we know they are simple arrays.