feat(lwc-template-compiler): Add new compiler API compileToFunction (#528)
## Details

This PR updates the `lwc-template-compiler` package to add `compileToFunction` API. This method takes as parameter a string template and a configuration and returns the instantiated template function.

The primary use-case of this new API is to allow inline compilation of template for unit tests. This PR also introduce the `compileTemplate` test utility to the `lwc-engine` package unit tests. Once this utility is stable, we will move it to the official `lwc-tests-utils` package for the general consumption.

## Does this PR introduce a breaking change?

* [ ] Yes
* [X] No