perf(events): reduce polymorphism of event listener objects
This fix improves the `tablecmp-create-10k.benchmark` benchmark in IE11 by roughly ~130ms by avoiding polymorphism on both the listener `function` object and the `VNode` object.

Setting a property on a `function` is polymorphic, so we avoid it. Adding properties to the `VNode` object is also polymorphic, so we avoid it. In the second case, we use `VNode.data` which is already a property bag designed for arbitrary third-party data.