perf: microoptimization of global markers
function calls:

before in each call to start/endGlobalMeasure the condition !isUserTimingSupported was being executed. But isUserTimingSupported is calculated once when importing the module.

Now when isUserTimingSupported is calculated if is true we export the start/endGlobalMeasure without the condition, and if ! isUserTimingSupported, we export a noop function as start/endGlobalMeasure