feat(engine): performance marks (#756)
* feat(engine): performance marks

Adds performance marks for:

- createElement
- render/patch generated from connecting a root element.
- flush rehydrate queue.

* wip: address pr comments

- enum
- init instead of createElement
- hydrate instead of connectedRootElement.

* wip: add measures for wc

entry point.

* perf: microoptimization of global markers

function calls:

before in each call to start/endGlobalMeasure the condition !isUserTimingSupported was being executed. But isUserTimingSupported is calculated once when importing the module.

Now when isUserTimingSupported is calculated if is true we export the start/endGlobalMeasure without the condition, and if ! isUserTimingSupported, we export a noop function as start/endGlobalMeasure

* fix: incorrect marker placement on wc.ts

The markers were added in the incorrect place when building a wc.