fix(css-compiler): remove unknown attribute without a tag
Removes the validation rule: unknown attribute without a tag.

Reason:

Before: if you have a rule .foo[disabled]: {}, the css compiler would throw with message Invalid usage of attribute selector disabled. For validation purposes, attributes that are not global attributes must be associated with a tag name when used in a CSS selector.

This rule can be valid for simplicity purposes, the disabled attribute is valid in button, fieldset, input etc. thus having this one rule makes sense, with the restriction you would need 3 (or more) rules