fix(synthetic-shadow): fix MutationObserver disconnect() memory leak
When a MutationObserver observes a DOM node, it creates a link from that DOM node to itself. When `disconnect()` is called, it should remove the association to avoid leaking the MutationObserver itself.

This is especially important when observing DOM nodes like `document.body` which are never removed from the DOM.