fix: slotted content of slotted custom element within if not visible (#1787) (#1794)
When passing slotted content to a custom element which is also slotted and wrapped in a if:true={foo} directive and toggle the if multiple times, the passed slot content stops being visible after the second toggle. You can see a repro here: https://developer.salesforce.com/docs/component-library/tools/playground/mFcLmzSYw/37/edit

The main issue is that the vnode for the slotted custom element is being reused, and during slot allocation the first time, vnode.children is being emptied.