fix(lwc): add missing `LightningElement.CustomElementConstructor` type
Currently, `create-lwc-app` [generates the following code inside
`src/index.ts`][1] when using typescript:

```typescript
import { createElement } from 'lwc';
import MyApp from 'my/app';

// ... redacted for brevity

createElement('my-app', { is: MyApp });
```

However, since there is no type for `createElement`, developers will see
errors in their editors from Typescript indicating that the `'lwc'`
package does not export a `createElement` function. Someone [previously tried to
add types for `createElement` in a pull request][2], but from reading
the PR's comments, this is a [feature, not a bug][3]. `createElement` is
a privileged function that only framework internals/special consumers
should be calling.

`create-lwc-app` should be updated to generate the following code
(suggested by [@caridy][4]) that is fully in the public API as it uses
web standards to define custom elements instead:

```typescript
import MyComponent from './some/path';
customElements.define('x-the-tagname-for-my-component', MyComponent.CustomElementConstructor);
const elm = document.createElement('x-the-tagname-for-my-component');
```

This commit adds typing for the static `CustomElementConstructor`
available on `LightningElement` and any subclasses as a pre-requisite
for updating the template in `create-lwc-app`.

[1]: https://github.com/muenzpraeger/create-lwc-app/blob/7fd23520b6d1b4b430b8ea66d3d36f1597e50238/packages/create-lwc-app/templates/src/client/index.ts#L50
[2]: https://github.com/salesforce/lwc/pull/1991/files
[3]: https://github.com/salesforce/lwc/pull/1991/files#r461293331
[4]: https://github.com/caridy