test(snapshot): use Jest snapshots for some tests
This PR migrates some tests to use [Jest snapshots](https://jestjs.io/docs/snapshot-testing). This makes it easier to update these tests, as you can just run `yarn test -u` and the snapshot files update automatically.

The downside of this PR is that the "expected" output files are now in Jest snapshot format rather than plaintext format. This may make it a bit harder to edit or inspect manually.