fix: use patched `parentNode` to navigate up while finding nearest root
We were using native `parentNode` while navigating up to find the nearest root node.
This made sense when the element was inserted via `lwc:dom="manual"`.

It breaks down when an element is slotted into a Shadow DOM element, because the
native `parentNode` doesn't give us the right parent node (it gives us the slot).