chore(types): fiddle with types
I enabled `isolatedDeclarations` and fixed everything that made sense.
Fully enabling the flag requires creating explicit types for the error info objects in `@lwc/errors`,
which would be tedious, so I skipped that and disabled the flag.

I enabled `isolatedModules` and fixed everything that made sense.
We can't use `const enum` with isolated modules, but we like them because they minify cleanly,
so I skipped that and disabled the flag.

I enabled `erasableSyntaxOnly` and fixed everything that made sense.
We can't use `const enum` with isolated modules, but we like them because they minify cleanly,
so I skipped that and disabled the flag.
(I did replace our regular `enum` with objects.)