feat(wire): make type def more helpful but less complete
Previously, `@wire(adapter,{config:value})` with an invalid `value`
resulted in a confusing error reported on `adapter`, because the function
inferred the config type and derived the adapter type.
Now the function infers the adapter type and derives the config type.
As a result, an invalid `value` correctly reports an error on `value`.
It does so by augmenting the config with the $reactive prop strings that
are available. However, due to recursive types, this list is potentially
infinite. To provide utility but mitigate this, only top-level props are
validated. Any chained property access (`$nested.prop`) validates that
the top-level prop is an object, but allows any nested props, even if
they are actually invalid.